feat: 이메일 패스워드 로그인 관련 기능 삭제 #18
