+++ lib/auth.dart
@@ -1,5 +1,5 @@
 import 'package:calendar_app/home.dart';
-import 'package:calendar_app/login.dart';
+import 'package:calendar_app/authContainer.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 
@@ -19,7 +19,7 @@ class AuthPage extends StatelessWidget {
 
           // user not logged in
           else {
-            return const Login();
+            return const AuthContainer();
           }
         }),
       ),
